backend/requirements.txt
flask>=3.0.3,<4
flask-cors>=4.0.1,<5
scikit-learn>=1.4.2
pandas>=2.2.0
numpy>=1.26.0,<3
joblib>=1.4.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
